requirements.txt
discord.py
deep-translator
translate
franco_arabic_transliterator